Refactor favorites into full stack component
