+++ app/routes/blog_/$slug.tsx
@@ -7,7 +7,6 @@ import { BlurrableImage } from '#app/components/blurrable-image.tsx'
 import { CourseCard } from '#app/components/course-card.tsx'
 import { GeneralErrorBoundary } from '#app/components/error-boundary.tsx'
 import { FourHundred, FourOhFour } from '#app/components/errors.tsx'
-import { FavoriteToggle } from '#app/components/favorite-toggle.tsx'
 import { Grid } from '#app/components/grid.tsx'
 import { BlogSection } from '#app/components/sections/blog-section.tsx'
 import { Spacer } from '#app/components/spacer.tsx'
@@ -15,6 +14,7 @@ import { TeamStats } from '#app/components/team-stats.tsx'
 import { H2, H4, H6, Paragraph } from '#app/components/typography.tsx'
 import { externalLinks } from '#app/external-links.tsx'
 import { getImageBuilder, getImgProps, images } from '#app/images.tsx'
+import { FavoriteToggle } from '#app/routes/resources/favorite.tsx'
 import { type KCDHandle, type MdxListItem, type Team } from '#app/types.ts'
 import {
 	getBlogReadRankings,
